requirements.txt
openenv-core[core]>=0.2.2
openai>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
